+++ packages/example/pages/_app.tsx
@@ -159,6 +159,9 @@ type RadiusScale = typeof radiusScales[number];
 const overlayBlurs = ['large', 'small', 'none'] as const;
 type OverlayBlur = typeof overlayBlurs[number];
 
+const modalSizes = ['wide', 'compact'] as const;
+type ModalSize = typeof modalSizes[number];
+
 function RainbowKitApp({ Component, pageProps }: AppProps) {
   const { disconnect } = useDisconnect();
   const [selectedInitialChainId, setInitialChainId] = useState<number>();
@@ -170,6 +173,7 @@ function RainbowKitApp({ Component, pageProps }: AppProps) {
   const [authEnabled, setAuthEnabled] = useState(pageProps.session !== null);
   const [showRecentTransactions, setShowRecentTransactions] = useState(true);
   const [coolModeEnabled, setCoolModeEnabled] = useState(false);
+  const [modalSize, setModalSize] = useState<ModalSize>('wide');
   const [showDisclaimer, setShowDisclaimer] = useState(false);
   const [customAvatar, setCustomAvatar] = useState(false);
 
@@ -204,6 +208,7 @@ function RainbowKitApp({ Component, pageProps }: AppProps) {
         chains={chains}
         coolMode={coolModeEnabled}
         initialChain={selectedInitialChainId}
+        modalSize={modalSize}
         showRecentTransactions={showRecentTransactions}
         theme={currentTheme({
           ...accentColor,
@@ -332,6 +337,23 @@ function RainbowKitApp({ Component, pageProps }: AppProps) {
                           onChange={e => setCustomAvatar(e.target.checked)}
                           type="checkbox"
                         />
+                      </td>
+                    </tr>
+                    <tr>
+                      <td>modalSize</td>
+                      <td>
+                        <select
+                          onChange={e =>
+                            setModalSize(e.target.value as ModalSize)
+                          }
+                          value={modalSize}
+                        >
+                          {modalSizes.map(size => (
+                            <option key={size} value={size}>
+                              {size}
+                            </option>
+                          ))}
+                        </select>
                       </td>
                     </tr>
                     <tr>