+++ site/lib/docsRoutes.ts
@@ -22,6 +22,7 @@ export const docsRoutes: RouteProps[] = [
     pages: [
       { title: 'Installation', slug: 'installation' },
       { title: 'ConnectButton', slug: 'connect-button' },
+      { title: 'Modal Sizes', slug: 'modal-sizes' },
       { title: 'Theming', slug: 'theming' },
       { title: 'Chains', slug: 'chains' },
       { title: 'Authentication', slug: 'authentication' },