+++ examples/with-next/CHANGELOG.md
@@ -1,5 +1,12 @@
 # with-next
 
+## 0.0.17
+
+### Patch Changes
+
+- Updated dependencies [1a4f2f7]
+  - @rainbow-me/rainbowkit@0.4.7
+
 ## 0.0.16
 
 ### Patch Changes